push

+++ src/main/java/ru/vadim/spring/dao/UserDao.java
@@ -7,12 +7,12 @@ public interface UserDao {
 
    List<User> allUsers();
 
-   User showUserById(int id);
+   User showUserById(Long id);
 
    void saveUser(User user);
 
-   void updateUser(int id, User user);
+   void updateUser(Long id, User user);
 
-   void deleteUser(int id);
+   void deleteUser(Long id);
 
 }
